+++ src/biodwh2-hpo/src/main/java/de/unibi/agbi/biodwh2/hpo/model/PhenotypeToGenesEntry.java
@@ -3,23 +3,14 @@ package de.unibi.agbi.biodwh2.hpo.model;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
-@JsonPropertyOrder({
-        "HPO-id", "HPO label", "entrez-gene-id", "entrez-gene-symbol", "Additional Info from G-D source", "G-D source",
-        "disease-ID for link"
-})
+@JsonPropertyOrder({"hpo_id", "hpo_name", "ncbi_gene_id", "gene_symbol"})
 public final class PhenotypeToGenesEntry {
-    @JsonProperty("HPO-id")
+    @JsonProperty("hpo_id")
     public String hpoId;
-    @JsonProperty("HPO label")
-    public String hpoLabel;
-    @JsonProperty("entrez-gene-id")
-    public Integer entrezGeneId;
-    @JsonProperty("entrez-gene-symbol")
-    public String entrezGeneSymbol;
-    @JsonProperty("Additional Info from G-D source")
-    public String additionalInfoFromGDSource;
-    @JsonProperty("G-D source")
-    public String gdSource;
-    @JsonProperty("disease-ID for link")
-    public String diseaseIdForLink;
+    @JsonProperty("hpo_name")
+    public String hpoName;
+    @JsonProperty("ncbi_gene_id")
+    public Integer ncbiGeneId;
+    @JsonProperty("gene_symbol")
+    public String geneSymbol;
 }